Property to set project id so that multiple SonarQube projects can comment on the same pull request.

+++ src/main/java/org/sonar/plugins/github/PullRequestFacade.java
@@ -35,6 +35,8 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import javax.annotation.CheckForNull;
 import javax.annotation.Nullable;
+
+import org.apache.commons.lang.StringUtils;
 import org.kohsuke.github.GHCommitState;
 import org.kohsuke.github.GHCommitStatus;
 import org.kohsuke.github.GHIssueComment;
@@ -76,6 +78,7 @@ public class PullRequestFacade {
 
   public PullRequestFacade(GitHubPluginConfiguration config) {
     this.config = config;
+
   }
 
   public void init(int pullRequestNumber, File projectBaseDir) {
@@ -138,6 +141,10 @@ public class PullRequestFacade {
     for (GHPullRequestReviewComment comment : pr.listReviewComments()) {
       if (!myself.equals(comment.getUser().getLogin())) {
         // Ignore comments from other users
+        continue;
+      }
+      if (!StringUtils.isEmpty(config.projectId()) && !comment.getBody().startsWith(MarkDownUtils.projectId(config.projectId()))) {
+        // Ignore comments that don't contain projectId
         continue;
       }
       if (!existingReviewCommentsByLocationByFile.containsKey(comment.getPath())) {
@@ -259,6 +266,10 @@ public class PullRequestFacade {
     boolean found = false;
     for (GHIssueComment comment : pr.listComments()) {
       if (myself.equals(comment.getUser().getLogin())) {
+        if (!StringUtils.isEmpty(config.projectId()) && !comment.getBody().startsWith(MarkDownUtils.projectId(config.projectId()))) {
+          // Ignore comments that don't contain projectId
+          continue;
+        }
         if (markup == null || found || !markup.equals(comment.getBody())) {
           comment.delete();
           continue;
